refactor: move controllers to v1, separate exposed api and authed api, change @Autowired to constructor injection
