+++ src/main/java/com/nivlalulu/nnpro/controller/v1/ProductController.java
@@ -1,9 +1,10 @@
-package com.nivlalulu.nnpro.controller;
+package com.nivlalulu.nnpro.controller.v1;
 
 import com.nivlalulu.nnpro.dto.v1.ProductDto;
 import com.nivlalulu.nnpro.dto.ApiResponse;
 import com.nivlalulu.nnpro.service.impl.ProductService;
 import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.annotation.Validated;
@@ -14,11 +15,10 @@ import java.util.UUID;
 import java.util.stream.Collectors;
 
 @RestController
-@RequestMapping("/product")
+@RequestMapping("/v1/product")
+@RequiredArgsConstructor
 @Validated
 public class ProductController {
-
-    @Autowired
     private ProductService productService;
 
     @PostMapping("/saveProduct")
